Set up card action panel
