+++ Assets/Plugins/FinolDigital.Cgs.CardGameDef.dll
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:fd99a4d3e6b0054dab4347fa0da18a055af6f69c277a76c6ac2a8aba5654cf7e
+oid sha256:0a50ea9a9a8b7b46f0cbaee2aac97a697ff4743ea49c09aba83010a864e6094e
 size 73216
