+++ Assets/Prefabs/CardGameView/Viewer/Card Action Panel.prefab.meta
