+++ Assets/Scripts/Cgs/CardGameView/Viewer/CardActionPanel.cs
@@ -1,0 +1,165 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+using System.Collections.Generic;
+using Cgs.CardGameView.Multiplayer;
+using Cgs.Play.Multiplayer;
+using JetBrains.Annotations;
+using UnityEngine;
+using UnityEngine.UI;
+
+namespace Cgs.CardGameView.Viewer
+{
+    public delegate void CardAction(CardModel cardModel);
+
+    public class CardActionPanel : MonoBehaviour
+    {
+        public static IReadOnlyDictionary<FinolDigital.Cgs.CardGameDef.CardAction, CardAction> ActionsDictionary =>
+            _actionsDictionary ??= new Dictionary<FinolDigital.Cgs.CardGameDef.CardAction, CardAction>
+            {
+                [FinolDigital.Cgs.CardGameDef.CardAction.Move] = Move,
+                [FinolDigital.Cgs.CardGameDef.CardAction.Rotate] = Rotate,
+                [FinolDigital.Cgs.CardGameDef.CardAction.Tap] = Tap,
+                [FinolDigital.Cgs.CardGameDef.CardAction.Flip] = Flip,
+                [FinolDigital.Cgs.CardGameDef.CardAction.Discard] = Discard
+            };
+
+        private static Dictionary<FinolDigital.Cgs.CardGameDef.CardAction, CardAction> _actionsDictionary;
+
+        public Button moveButton;
+        public Button rotateButton;
+        public Button tapButton;
+        public Button flipButton;
+        public Button discardButton;
+
+        public void Show(Vector2 position)
+        {
+            gameObject.SetActive(true);
+
+            transform.position = position;
+
+            rotateButton.interactable =
+                CardViewer.Instance.SelectedCardModel.ParentCardZone != null &&
+                CardViewer.Instance.SelectedCardModel.ParentCardZone.allowsRotation;
+            rotateButton.transform.GetChild(0).GetComponent<Image>().color =
+                CardGameManager.Current.GameDefaultCardAction == FinolDigital.Cgs.CardGameDef.CardAction.Rotate
+                    ? Color.green
+                    : Color.white;
+
+            tapButton.interactable =
+                CardViewer.Instance.SelectedCardModel.ParentCardZone != null &&
+                CardViewer.Instance.SelectedCardModel.ParentCardZone.allowsRotation;
+            tapButton.transform.GetChild(0).GetComponent<Image>().color =
+                CardGameManager.Current.GameDefaultCardAction == FinolDigital.Cgs.CardGameDef.CardAction.Tap
+                    ? Color.green
+                    : Color.white;
+
+            flipButton.interactable =
+                CardViewer.Instance.SelectedCardModel.ParentCardZone != null &&
+                CardViewer.Instance.SelectedCardModel.ParentCardZone.allowsFlip;
+            flipButton.transform.GetChild(0).GetComponent<Image>().color =
+                CardGameManager.Current.GameDefaultCardAction == FinolDigital.Cgs.CardGameDef.CardAction.Flip
+                    ? Color.green
+                    : Color.white;
+        }
+
+        public void Update()
+        {
+            if (Inputs.IsNew && moveButton.interactable)
+                Move(CardViewer.Instance.SelectedCardModel);
+            else if (Inputs.IsLoad && rotateButton.interactable)
+                Rotate(CardViewer.Instance.SelectedCardModel);
+            else if (Inputs.IsSave && tapButton.interactable)
+                Tap(CardViewer.Instance.SelectedCardModel);
+            else if (Inputs.IsFilter && flipButton.interactable)
+                Flip(CardViewer.Instance.SelectedCardModel);
+            else if (Inputs.IsOption && discardButton.interactable)
+                Discard(CardViewer.Instance.SelectedCardModel);
+        }
+
+        [UsedImplicitly]
+        public void Move()
+        {
+            Move(CardViewer.Instance.SelectedCardModel);
+        }
+
+        public static void Move(CardModel cardModel)
+        {
+            CardGameManager.Instance.Messenger.Show("Move Feature is Coming Soon!");
+        }
+
+        [UsedImplicitly]
+        public void Rotate()
+        {
+            Rotate(CardViewer.Instance.SelectedCardModel);
+        }
+
+        public static void Rotate(CardModel cardModel)
+        {
+            if (cardModel.ParentCardZone == null || !cardModel.ParentCardZone.allowsRotation)
+            {
+                Debug.LogWarning("Ignoring rotation request since the parent card zone does not support it.");
+                return;
+            }
+
+            cardModel.Rotation *= Quaternion.Euler(0, 0, -CardGameManager.Current.GameCardRotationDegrees);
+        }
+
+        [UsedImplicitly]
+        public void Tap()
+        {
+            Tap(CardViewer.Instance.SelectedCardModel);
+        }
+
+        public static void Tap(CardModel cardModel)
+        {
+            if (cardModel.ParentCardZone == null || !cardModel.ParentCardZone.allowsRotation)
+            {
+                Debug.LogWarning("Ignoring rotation request since the parent card zone does not support it.");
+                return;
+            }
+
+            var unTappedRotation = Quaternion.identity;
+            if (CgsNetManager.Instance != null && CgsNetManager.Instance.LocalPlayer != null)
+                unTappedRotation = CgsNetManager.Instance.LocalPlayer.DefaultRotation;
+            var isTapped = !unTappedRotation.Equals(cardModel.Rotation);
+            var tappedRotation = unTappedRotation *
+                                 Quaternion.Euler(0, 0, -CardGameManager.Current.GameCardRotationDegrees);
+            cardModel.Rotation = isTapped ? unTappedRotation : tappedRotation;
+        }
+
+        [UsedImplicitly]
+        public void Flip()
+        {
+            Flip(CardViewer.Instance.SelectedCardModel);
+        }
+
+        public static void Flip(CardModel cardModel)
+        {
+            if (cardModel.ParentCardZone == null || !cardModel.ParentCardZone.allowsFlip)
+            {
+                Debug.LogWarning("Ignoring flip request since the parent card zone does not support it.");
+                return;
+            }
+
+            cardModel.IsFacedown = !cardModel.IsFacedown;
+        }
+
+        [UsedImplicitly]
+        public void Discard()
+        {
+            Discard(CardViewer.Instance.SelectedCardModel);
+        }
+
+        public static void Discard(CardModel cardModel)
+        {
+            cardModel.Discard();
+        }
+
+        public void Hide()
+        {
+            gameObject.SetActive(false);
+        }
+    }
+}
