+++ Assets/Scripts/Cgs/Play/PlayController.cs
@@ -144,7 +144,6 @@ namespace Cgs.Play
         {
             Instantiate(cardViewerPrefab);
             Instantiate(playableViewerPrefab);
-            CardViewer.Instance.GetComponent<CardActions>().Show();
             if (Menu.Settings.PreviewOnMouseOver)
                 CardViewer.Instance.Mode = CardViewerMode.Expanded;
             CardGameManager.Instance.OnSceneActions.Add(ResetPlayArea);
@@ -415,7 +414,7 @@ namespace Cgs.Play
 
         public static void SetPlayActions(CardModel cardModel)
         {
-            cardModel.DefaultAction = CardActions.ActionsDictionary[CardGameManager.Current.GameDefaultCardAction];
+            cardModel.DefaultAction = CardActionPanel.ActionsDictionary[CardGameManager.Current.GameDefaultCardAction];
             cardModel.SecondaryDragAction = cardModel.Rotate;
         }
 
@@ -592,7 +591,7 @@ namespace Cgs.Play
                 return;
 
             cardModel.SecondaryDragAction = cardModel.UpdateParentCardZoneScrollRect;
-            cardModel.DefaultAction = CardActions.ActionsDictionary[CardGameManager.Current.GameDefaultCardAction];
+            cardModel.DefaultAction = CardActionPanel.ActionsDictionary[CardGameManager.Current.GameDefaultCardAction];
         }
 
         public static void OnAddCardModelFaceDown(CardZone cardZone, CardModel cardModel)
@@ -601,7 +600,7 @@ namespace Cgs.Play
                 return;
 
             cardModel.SecondaryDragAction = cardModel.UpdateParentCardZoneScrollRect;
-            cardModel.DefaultAction = CardActions.ActionsDictionary[CardGameManager.Current.GameDefaultCardAction];
+            cardModel.DefaultAction = CardActionPanel.ActionsDictionary[CardGameManager.Current.GameDefaultCardAction];
             cardModel.IsFacedown = true;
         }
 
@@ -611,7 +610,7 @@ namespace Cgs.Play
                 return;
 
             cardModel.SecondaryDragAction = cardModel.UpdateParentCardZoneScrollRect;
-            cardModel.DefaultAction = CardActions.ActionsDictionary[CardGameManager.Current.GameDefaultCardAction];
+            cardModel.DefaultAction = CardActionPanel.ActionsDictionary[CardGameManager.Current.GameDefaultCardAction];
             cardModel.IsFacedown = false;
         }
